PHP5.4 compatibility.  String offsets must not only be integral, they must also have integer type.

+++ includes/functions/functions_print_lists.php
@@ -2012,11 +2012,11 @@ function print_chart_by_decade($data, $title) {
 	$chart_url .= "&amp;chd=s:"; // data : simple encoding from A=0 to 9=61
 	$CHART_ENCODING61 = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
 	for ($y=1570; $y<2030; $y+=10) {
-		$chart_url .= $CHART_ENCODING61[floor(substr_count($data[$y], "M")*61/$vmax)];
+		$chart_url .= $CHART_ENCODING61[(int)floor(substr_count($data[$y], "M")*61/$vmax)];
 	}
 	$chart_url .= ",";
 	for ($y=1570; $y<2030; $y+=10) {
-		$chart_url .= $CHART_ENCODING61[floor(substr_count($data[$y], "F")*61/$vmax)];
+		$chart_url .= $CHART_ENCODING61[(int)floor(substr_count($data[$y], "F")*61/$vmax)];
 	}
 	$html = '<img src="'. $chart_url. '" alt="'. $title. '" title="'. $title. '" class="gchart">';
 	return $html;
